refactor: separate source to module
