+++ backend/src/main/java/io/github/engman/authentication/web/AuthenticationRequest.java
@@ -1,4 +1,4 @@
-package io.github.engman.web;
+package io.github.engman.authentication.web;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
