+++ backend/src/main/java/io/github/engman/user/UserRepository.java
@@ -1,10 +1,8 @@
-package io.github.engman.repository;
+package io.github.engman.user;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import io.github.engman.domain.User;
 
 public interface UserRepository extends JpaRepository<User, Long> {
